Move connector creation logic to common place
