+++ crates/router/src/core/payments.rs
@@ -1,4 +1,5 @@
 pub mod access_token;
+pub mod customers;
 pub mod flows;
 pub mod helpers;
 pub mod operations;
