+++ crates/router/src/types.rs
@@ -290,6 +290,7 @@ pub struct VerifyRequestData {
     pub setup_future_usage: Option<storage_enums::FutureUsage>,
     pub off_session: Option<bool>,
     pub setup_mandate_details: Option<payments::MandateData>,
+    pub email: Option<masking::Secret<String, Email>>,
 }
 
 #[derive(Debug, Clone)]
